backend/requirements.txt
# Testing requirements.txt with ML dependencies (CPU only)
# Core FastAPI and database
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
jwt==1.3.1

# File handling
python-multipart==0.0.20
aiofiles==24.1.0

# Validation and settings
pydantic[email]==2.11.3
pydantic-settings==2.9.1
email-validator==2.2.0
python-dotenv==1.1.0

# Azure integration
azure-storage-blob==12.25.1

# HTTP and utilities
requests==2.28.2
httpx==0.13.3
httpcore==0.9.1
Jinja2==3.1.6

# Audio processing
SpeechRecognition==3.14.3
googletrans==4.0.0rc1
gTTS==2.5.4
pydub==0.25.1

# Data processing
pandas==2.2.3
numpy==1.26.4
tqdm==4.65.2

# ML frameworks (CPU versions)
torch==2.0.0
torchvision==0.15.1

# Computer vision
opencv-python==4.11.0.86
pillow==10.4.0

# Video processing
moviepy==2.1.2
imageio==2.37.0
imageio-ffmpeg==0.6.0

# Plotting and science
matplotlib==3.10.1
seaborn==0.13.2
scipy==1.15.2

# MMPose ecosystem - try without exact versions to avoid compilation issues
mmcv>=2.0.0
mmengine>=0.10.0
mmdet>=3.0.0
mmpose>=1.0.0

# Utilities
decorator==5.1.1
proglog==0.1.12

# Development dependencies (optional for deployment)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
